Refactored profanity detection logic by introducing `ProfanityClassifier`, replaced inline implementations with reusable class across pipeline and tests, and removed redundant `containsProfanity` methods.
